+++ CHANGELOG.md
@@ -18,6 +18,8 @@ limitations under the License.
 
 ## Unreleased
 
+- fix: recover handling `KeyboardInterrupt` in `triton.serve()`
+
 [//]: <> (put here on external component update with short summary what change or link to changelog)
 
 - Version of [Triton Inference Server](https://github.com/triton-inference-server/) embedded in wheel: [2.36.0](https://github.com/triton-inference-server/server/releases/tag/v2.36.0)